Split semantic CSS into components
This converts the semantic ui css file into a big import for individual components and disables many that don't appear to be in use.

I've tried to identify these by searching through all templates for class references and then examining pages for reverts. Everything seems the same, but if something has regressed it can be re-enabled by uncommenting the relevant import.

All in all it seems have reduced the filesize by about 250k.