Replace structural divs with semantic elements
This commit splits the page up into header, footer, main and sections. It also replaces the main navigation div with an accesible nav and ul structure.

I'm not sure how the mobile menu will work with screen readers as orca is currently broken for me. I suspect duplicating the links without aria attributes etc is not ideal, but not sure of the options with semantic ui.

The extra css is to fix the display of the nav menu now it is a ul.