A11y improvements. (#402)
* Replace structural divs with semantic elements

This commit splits the page up into header, footer, main and sections. It also replaces the main navigation div with an accesible nav and ul structure.

I'm not sure how the mobile menu will work with screen readers as orca is currently broken for me. I suspect duplicating the links without aria attributes etc is not ideal, but not sure of the options with semantic ui.

The extra css is to fix the display of the nav menu now it is a ul.

* Refactor heading levels

This will result in a small amount of visual change (sub-page headings will be slightly bigger) but should give a more logical structure to the document in terms of heading levels.

* Make category items lists

Converts the list of items in each category to a unordered list. This should announce as a list to screen readers so will give those with poor / no vision more context to the page content.

* Fix colour contrast on footer link.

* Fix missing header styles.