Replace Shields.io with locally rendered data/badges (#591)
* Improve accessibility of badges

* Replace Shields.io with locally rendered data/badges

* Fix missing repo info when repository URL ends in a slash

* Fix missing activity info for items without source

* Fix missing repo info when repository URL ends in .git

* Format star count

* Fix GitHub icon displaying even when repo is on a different site

* More accessibility tweaks

* Add Gitea stats support

* Re-order card markup for better accessibility

I tried the page with NVDA, and the aria-label approach didn't work (apparently it's not valid for 'li' tags). So instead, I just tweaked the markup to properly present the item's title first.

* Add build date to the footer

* Hide license badge when GitHub check inconclusive

A lot of the cases where 'Other' was showing as the license turned out to be MIT/Apache projects with their files named wrong, which feels misleading.